Relocate cuboid_subvolume_labels into catalog_analysis_helpers
